+++ db/events.php
@@ -31,4 +31,10 @@ $observers = [
         'priority'    => 9999,
         'internal'    => false,
     ],
+    [
+        'eventname'   => '\core\event\course_restored',
+        'callback'    => '\local_navigatr\observer::course_restored',
+        'priority'    => 9999,
+        'internal'    => false,
+    ],
 ];